Extract UserButton into its own component and update stories
