+++ packages/studio-base/src/components/AppBar/UserMenu.tsx
@@ -85,9 +85,6 @@ export function UserMenu({
     window.open(process.env.FOXGLOVE_ACCOUNT_DASHBOARD_URL, "_blank");
   }, []);
 
-  if (currentUser == undefined) {
-    return <></>;
-  }
   return (
     <>
       <Menu
@@ -104,15 +101,19 @@ export function UserMenu({
         <MenuItem onClick={onPreferencesClick}>
           <ListItemText primary="Preferences" />
         </MenuItem>
-        <MenuItem onClick={onSettingsClick}>
-          <ListItemText primary={currentUser.email} />
-        </MenuItem>
-        <Divider variant="middle" />
-        <MenuItem onClick={onSignoutClick}>
-          <ListItemText>
-            <Typography color="error">Sign out</Typography>
-          </ListItemText>
-        </MenuItem>
+        {currentUser && (
+          <MenuItem onClick={onSettingsClick}>
+            <ListItemText primary={currentUser.email} />
+          </MenuItem>
+        )}
+        {currentUser && <Divider variant="middle" />}
+        {currentUser && (
+          <MenuItem onClick={onSignoutClick}>
+            <ListItemText>
+              <Typography color="error">Sign out</Typography>
+            </ListItemText>
+          </MenuItem>
+        )}
       </Menu>
       {confirmModal}
     </>
